+++ lib/fulfil/remote/resource.rb
@@ -37,7 +37,62 @@ module Fulfil
         other.respond_to?(:id) && id == other.id
       end
 
+      private
+
+      # Extracts an id from a Fulfil reference payload.
+      #
+      # Fulfil references are often returned as either:
+      # - Integer id
+      # - Array payload like [id, rec_name]
+      #
+      # @param value [Integer, Array, nil]
+      # @return [Integer, nil]
+      def extract_remote_id(value)
+        value.is_a?(Array) ? value.first : value
+      end
+
       class << self
+        # Defines a has_many association loaded by a list of remote ids.
+        #
+        # @param name [Symbol] association method name
+        # @param class_name [String] fully qualified class name
+        # @param ids_key [String, Symbol] local attribute key with id list
+        # @return [void]
+        # rubocop:disable Naming/PredicatePrefix
+        def has_many(name, class_name:, ids_key:)
+          define_method(name) do
+            ids = attributes[ids_key.to_s] || []
+            return [] if ids.empty?
+
+            self.class.constantize(class_name).all(ids: ids)
+          end
+        end
+        # rubocop:enable Naming/PredicatePrefix
+
+        # Defines a belongs_to association loaded from a Fulfil reference payload.
+        #
+        # @param name [Symbol] association method name
+        # @param class_name [String] fully qualified class name
+        # @param foreign_key [String, Symbol] local attribute key containing reference
+        # @return [void]
+        def belongs_to(name, class_name:, foreign_key:)
+          define_method(name) do
+            remote_reference = attributes[foreign_key.to_s]
+            remote_id = extract_remote_id(remote_reference)
+            return nil if remote_id.nil?
+
+            self.class.constantize(class_name).find(remote_id)
+          end
+        end
+
+        # Resolves a constant from a fully qualified class name.
+        #
+        # @param class_name [String]
+        # @return [Class]
+        def constantize(class_name)
+          class_name.split('::').reject(&:empty?).reduce(Object) { |scope, const| scope.const_get(const) }
+        end
+
         # Fetches resources from Fulfil.
         #
         # @param ids [Array<Integer>] optional id filter